vidc:1080p: fix for wrong frame tags in slice delivery mode
The core sends the wrong frame tag from the second slice onwards
for certain resolutions.The video encode session fails if the frame
tags dont match the values present in the transaction table. This
workaround will use the frame tag of the first slice for the
remaining slices. We need this workaround till video core fixes
the issue.

Change-Id: I36f2401a3b6d1ec2f9d482968f7d0946a7733572
CRs-fixed: 368678
Signed-off-by: Rajeshwar Kurapaty <rkurapat@codeaurora.org>